+++ packages/terra-divider/tests/jest/Divider.test.jsx
@@ -14,7 +14,7 @@ describe('Divider', () => {
 
   // Custom Prop Test
   it('it should pass in a custom prop', () => {
-    const wrapper = shallow(<Divider id="testDivider" />);
+    const wrapper = shallow(<Divider id="testDivider" text="custom-text" level={3} />);
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -46,7 +46,7 @@ describe('Divider', () => {
   });
 
   it('should spread a custom prop onto the hr divider', () => {
-    const wrapper = shallow(<Divider id="custom-id" />);
+    const wrapper = shallow(<Divider id="custom-id" text="custom-text" level={3} />);
 
     expect(wrapper.is('#custom-id')).toBe(true);
   });
@@ -55,6 +55,12 @@ describe('Divider', () => {
     const wrapper = shallow(<Divider id="custom-id" text="custom-text" />);
 
     expect(wrapper.is('#custom-id')).toBe(true);
+  });
+
+  it('should spread a custom prop onto the text with level divider', () => {
+    const wrapper = shallow(<Divider id="custom-id" text="custom-text" level={3} />);
+
+    expect(wrapper).toMatchSnapshot();
   });
 
   it('should apply the theme context className to the hr divider', () => {